Remove InputLoopService from CosmosDB container sample
- Removed InputLoopService.cs file
- Moved input loop logic directly to Program.cs using recommended pattern
- Uses host.StartAsync(), GetRequiredService<IMessageSession>(), and host.StopAsync()
- Sample builds and maintains same functionality
