+++ samples/cosmosdb/container/CosmosDB_3/Server/Program.cs
@@ -15,68 +15,80 @@ class Program
           builder.AddConsole();
       }).CreateLogger<Program>();
 
-
     public static async Task Main(string[] args)
     {
+        Console.Title = "Server";
         logger.LogInformation("Starting application...");
-        var host = CreateHostBuilder(args).Build();
-        await host.RunAsync();
+        
+        var builder = Host.CreateApplicationBuilder(args);
+
+        logger.LogInformation("Configuring NServiceBus...");
+
+        #region CosmosDBConfig
+        var endpointConfiguration = new EndpointConfiguration("Samples.CosmosDB.Container.Server");
+
+        var persistence = endpointConfiguration.UsePersistence<CosmosPersistence>();
+        
+        // Get connection string from environment variable, fallback to local emulator
+        var connection = Environment.GetEnvironmentVariable("COSMOS_CONNECTION_STRING") 
+            ?? @"AccountEndpoint=https://localhost:8081/;AccountKey=C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==";
+        
+        logger.LogInformation($"Using Cosmos DB connection: {(Environment.GetEnvironmentVariable("COSMOS_CONNECTION_STRING") != null ? "Azure service" : "Local emulator")}");
+        
+        persistence.DatabaseName("Samples.CosmosDB.Container");
+        var cosmosClient = new CosmosClient(connection);
+        persistence.CosmosClient(cosmosClient);
+        persistence.DefaultContainer("OrderSagaData", "/id");
+
+        #endregion
+
+        endpointConfiguration.UseTransport(new LearningTransport());
+        endpointConfiguration.UseSerialization<SystemJsonSerializer>();
+        endpointConfiguration.EnableInstallers();
+
+        #region ContainerInformationFromLogicalMessage
+        var transactionInformation = persistence.TransactionInformation();
+        transactionInformation.ExtractContainerInformationFromMessage<ShipOrder>(m =>
+        {
+            logger.LogInformation($"Message '{m.GetType().AssemblyQualifiedName}' destined to be handled by '{nameof(ShipOrderSaga)}' will use 'ShipOrderSagaData' container.");
+            return new ContainerInformation("ShipOrderSagaData", new PartitionKeyPath("/id"));
+        });
+        #endregion
+        #region ContainerInformationFromHeaders
+        transactionInformation.ExtractContainerInformationFromHeaders(headers =>
+        {
+            if (headers.TryGetValue(Headers.SagaType, out var sagaTypeHeader) && sagaTypeHeader.Contains(nameof(ShipOrderSaga)))
+            {
+                logger.LogInformation($"Message '{headers[Headers.EnclosedMessageTypes]}' destined to be handled by '{nameof(ShipOrderSaga)}' will use 'ShipOrderSagaData' container.");
+
+                return new ContainerInformation("ShipOrderSagaData", new PartitionKeyPath("/id"));
+            }
+            return null;
+        });
+        #endregion
+
+        cosmosClient.CreateDatabaseIfNotExistsAsync("Samples.CosmosDB.Container").Wait();
+        var database = cosmosClient.GetDatabase("Samples.CosmosDB.Container");
+        database.CreateContainerIfNotExistsAsync("ShipOrderSagaData", "/id").Wait();
+
+        builder.UseNServiceBus(endpointConfiguration);
+
+        var host = builder.Build();
+        await host.StartAsync();
+
+        logger.LogInformation("Server started successfully. Press any key to exit");
+
+        // Wait for input or termination
+        if (Console.IsInputRedirected)
+        {
+            // In container or background mode, wait for termination signal
+            await host.WaitForShutdownAsync();
+        }
+        else
+        {
+            // Interactive mode
+            Console.ReadKey();
+            await host.StopAsync();
+        }
     }
-
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-     Host.CreateDefaultBuilder(args)
-         .ConfigureServices((hostContext, services) =>
-         {
-             Console.Title = "Server";
-             services.AddLogging();
-
-         }).UseNServiceBus(x =>
-         {
-             logger.LogInformation("Configuring NServiceBus...");
-
-             #region CosmosDBConfig
-             var endpointConfiguration = new EndpointConfiguration("Samples.CosmosDB.Container.Server");
-
-             var persistence = endpointConfiguration.UsePersistence<CosmosPersistence>();
-             var connection = @"AccountEndpoint=https://localhost:8081/;AccountKey=C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==";
-             persistence.DatabaseName("Samples.CosmosDB.Container");
-             var cosmosClient = new CosmosClient(connection);
-             persistence.CosmosClient(cosmosClient);
-             persistence.DefaultContainer("OrderSagaData", "/id");
-
-             #endregion
-
-             endpointConfiguration.UseTransport(new LearningTransport());
-             endpointConfiguration.UseSerialization<SystemJsonSerializer>();
-             endpointConfiguration.EnableInstallers();
-
-             #region ContainerInformationFromLogicalMessage
-             var transactionInformation = persistence.TransactionInformation();
-             transactionInformation.ExtractContainerInformationFromMessage<ShipOrder>(m =>
-             {
-                 logger.LogInformation($"Message '{m.GetType().AssemblyQualifiedName}' destined to be handled by '{nameof(ShipOrderSaga)}' will use 'ShipOrderSagaData' container.");
-                 return new ContainerInformation("ShipOrderSagaData", new PartitionKeyPath("/id"));
-             });
-             #endregion
-             #region ContainerInformationFromHeaders
-             transactionInformation.ExtractContainerInformationFromHeaders(headers =>
-             {
-                 if (headers.TryGetValue(Headers.SagaType, out var sagaTypeHeader) && sagaTypeHeader.Contains(nameof(ShipOrderSaga)))
-                 {
-                     logger.LogInformation($"Message '{headers[Headers.EnclosedMessageTypes]}' destined to be handled by '{nameof(ShipOrderSaga)}' will use 'ShipOrderSagaData' container.");
-
-                     return new ContainerInformation("ShipOrderSagaData", new PartitionKeyPath("/id"));
-                 }
-                 return null;
-             });
-             #endregion
-
-             cosmosClient.CreateDatabaseIfNotExistsAsync("Samples.CosmosDB.Container");
-             var database = cosmosClient.GetDatabase("Samples.CosmosDB.Container");
-             database.CreateContainerIfNotExistsAsync("ShipOrderSagaData", "/id");
-
-             Console.ReadKey();
-
-             return endpointConfiguration;
-         });
 }
